Completed Load & Verify for Presentations
There's a lot of small tweaks in a lot of different places in the code.
I'm passing the csvtype, set as a number 1 - 4, from load to verify, and
then using that to determine what kind of data we're interacting with.
It should be pretty simple for other people doing load/verify to just
copy my extra functions and add their own if statements for the csv
they'll be using.