Made Progress on Analyze for Teaching
Analyze for teaching now forms multiple trees, and takes note of
multiple values.

- Still missing the Other tree. When we want to check for data in
"Other" programs, we just need to send some arbitrary string to the
populate_for_teaching function, then check for it at the start and if
it's that string, we parse the data based on it NOT belonging to one of
the three specific programs.

- Still missing students column. I put a replacement column of
hours/session in for now. I think students column is supposed to be
total number of students but I'm not sure and I'm tired and I didn't
want to mess about figuring out how that was going to work. With this
code as is, it will work in the future (in theory).

- Still needs the graph stuff.