Analyze Teaching Bugfixes
Fixed the Filter bug, and a bug that was adding previous numbers from
one program into another (basically, the totals from UME stacked with
totals from PME, etc). This was just me not resetting tCount at the end
of the tree in analyze.

For the Filter bug - it was an error when filtering and one of the trees
is 0. I just added a check in tree_list_vo to ignore that tree if it's
empty.